requirements.txt
Flask==3.0.0
qrcode[pil]==7.4.2
opencv-python-headless==4.8.1.78
pyzbar==0.1.9
Pillow==10.1.0
numpy==1.26.2
